Declare fields in PatchscopeBase
Also moves the Context classes to patchscope base to avoid circular dependency
